Snapshot persistence
Added the ability to not have to redownload snapshots if they're already loaded

Next look into storing multiple snapshots locally since they're so small.

+++ scripts/build/build-rust.sh
@@ -94,10 +94,20 @@ SNAP_TARBALL=$(echo $SNAP_TARBALL | tr -s ' ' | cut -d ' ' -f 3)
 
 # setup snapshot
 cd $SNAP_DIR
-rm -rf *
-$DROPBOX -p download snapshots/$SNAP_TARBALL
-tar xjf $SNAP_TARBALL --strip-components=1
-rm $SNAP_TARBALL
+# Only need to download if our current snapshot is not at the right version
+INSTALLED_SNAPSHOT_VERSION=
+if [ -f VERSION ]; then
+  INSTALLED_SNAPSHOT_VERSION=$(cat VERSION)
+fi
+if [ "$SNAP_TARBALL" != "$INSTALLED_SNAPSHOT_VERSION" ]; then
+  rm -rf *
+  $DROPBOX -p download snapshots/$SNAP_TARBALL
+  tar xjf $SNAP_TARBALL --strip-components=1
+  rm $SNAP_TARBALL
+  echo "$SNAP_TARBALL" > VERSION
+else
+  echo "Requested snapshot $SNAP_TARBALL already installed, no need to re-download and install."
+fi
 bin/rustc -V
 
 # Get information about HEAD
@@ -184,7 +194,7 @@ fi
 
 # cleanup
 rm -rf $DIST_DIR/*
-rm -rf $SNAP_DIR/*
+#rm -rf $SNAP_DIR/*
 
 end=$(date +"%s")
 test_diff=$(($end-$starttest))
